refactor: Pass name of detector result class to FormattedOutputManager.
