+++ src/main/java/com/synopsys/integration/detect/workflow/report/output/FormattedOutputManager.java
@@ -41,6 +41,7 @@ import com.synopsys.integration.detect.workflow.result.DetectResult;
 import com.synopsys.integration.detect.workflow.status.DetectIssue;
 import com.synopsys.integration.detect.workflow.status.Status;
 import com.synopsys.integration.detect.workflow.status.UnrecognizedPaths;
+import com.synopsys.integration.detectable.Extraction;
 import com.synopsys.integration.detector.base.DetectorEvaluation;
 import com.synopsys.integration.detector.base.DetectorEvaluationTree;
 import com.synopsys.integration.util.NameVersion;
@@ -115,7 +116,7 @@ public class FormattedOutputManager {
         detectorOutput.extracted = evaluation.wasExtractionSuccessful();
         detectorOutput.discoverable = evaluation.wasDiscoverySuccessful();
         detectorOutput.status = evaluation.getStatus().name();
-        detectorOutput.statusCode = evaluation.getStatusCode();
+        detectorOutput.statusCode = formatResultNameToStatusCode(evaluation.getResultClassName());
         detectorOutput.statusReason = evaluation.getStatusReason();
 
         if (evaluation.getDiscovery() != null) {
@@ -132,6 +133,15 @@ public class FormattedOutputManager {
         }
 
         return detectorOutput;
+    }
+
+    private String formatResultNameToStatusCode(String resultName) {
+        String[] classnamePieces = resultName.split("\\.");
+        String actualResultName = classnamePieces[classnamePieces.length-1].replace("DetectResult", "").replace("DetectorResult", "").replace("DetectableResult", "");
+
+        return actualResultName
+                   .replaceAll("([a-z])([A-Z]+)", "$1_$2")
+                   .toUpperCase();
     }
 
     private void detectorsComplete(final DetectorToolResult detectorToolResult) {
